Issue #28: Support of help on YANG keywords and types in hover.
